Changed GraphicsProfile to Reach (instead of HiDef) to work on graphic cards without DirectX10.

+++ MyGame/MyGame/code/Render & Effects/TEX.cs
@@ -10,7 +10,7 @@ namespace MyGame
 {
     public class TEX
     {
-        public static int[] index = { 2, 0, 1, 3, 2, 1 };
+        public static short[] index = { 2, 0, 1, 3, 2, 1 };
 
         public Texture2D texture;
         public VertexPositionColorTexture[] vertex;
